Update index.html
The first page of the website, this is the basic layout of the webpage.

We set up the structure of how the site will look with a header, content and footer. All information is currently inserted via javascript.

The script source has to be at the bottom of the page for the javascript to work.